Add hipStreamGetCtx to get the context associated with a stream
Even tough the context API is deprecated, this is the only portable way
to query the device from a stream.
This is useful if for example a library wants to enable device specific
optimizations transparently.

Below is the example code to query the device from a stream wich is
possible with this change.
```c
hipCtx_t ctx;
hipStreamGetCtx(stream, &ctx);
hipCtxPushCurrent(ctx);

hipDevice_t device;
hipCtxGetDevice(&device);
hipCtxPopCurrent(nullptr);
```